+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+
+- Fixed: `peerDependencies` have been adjusted to minimal requirements
+
 # 0.1.0 - 2015-10-21
 
 💥