+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+- Fixed: collection is not limited to 8~10 items anymore
 - Fixed: `peerDependencies` have been adjusted to minimal requirements
 - Added: `MetadataProvider` to send the data you want to the React context
 in a `metadata` attribute.