+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fixed: collection is not limited to 8~10 items anymore
 - Fixed: `peerDependencies` have been adjusted to minimal requirements
 - Fixed: fetched `collection.json` now rely on the base url path
+- Removed: useless whitespace in `collection.json`
 - Changed: path(name) of the base url now have a trailing slash
 - Added: `MetadataProvider` to send the data you want to the React context
 in a `metadata` attribute.