+++ CHANGELOG.md
@@ -9,6 +9,8 @@ in a `metadata` attribute.
 - Added: `__BASE_URL__` is now send as constantes.
 - Added: `statinamic/lib/prepare-defined-values` helper to easily use JS
 object with `webpack.DefinePlugin`
+- Added: `defaultHead` option to `markdown-as-json-loader` loader to provide
+default values
 
 # 0.1.0 - 2015-10-21